WIP: retrieval almost working again
