+++ crates/learner/tests/workflows/paper_retrieval.rs
@@ -1,26 +1,32 @@
 use std::fs;
 
-use learner::resource::ResourceConfig;
+use learner::{
+  configuration::{Config, ConfigurationManager},
+  resource::ResourceTemplate,
+};
 
 use super::*;
 
 #[traced_test]
 #[tokio::test]
 async fn test_arxiv_retriever_integration() -> TestResult<()> {
-  let ret_config_str = fs::read_to_string("config/retrievers/arxiv.toml").expect(
-    "Failed to read config
-    file",
-  );
-  let res_config_str = fs::read_to_string("config/resources/paper.toml").expect(
-    "Failed to read config
-    file",
-  );
+  // let ret_config_str = fs::read_to_string("config/retrievers/arxiv.toml").expect(
+  //   "Failed to read config
+  //   file",
+  // );
+  // let res_config_str = fs::read_to_string("config/resources/paper.toml").expect(
+  //   "Failed to read config
+  //   file",
+  // );
+  let mut manager = ConfigurationManager::new(PathBuf::from("config_new"));
+  let retriever: Config<Retriever> = manager.load_config("config_new/arxiv.toml")?;
 
-  let retriever: Retriever = toml::from_str(&ret_config_str).expect("Failed to parse config");
-  let resource: ResourceConfig = toml::from_str(&res_config_str).expect("Failed to parse config");
+  // let retriever: Retriever = toml::from_str(&ret_config_str).expect("Failed to parse config");
+  // let resource: ResourceTemplate = toml::from_str(&res_config_str).expect("Failed to parse
+  // config");
 
   // Test with a real arXiv paper
-  let paper = retriever.retrieve_resource("2301.07041", &resource).await?;
+  let paper = retriever.inner().retrieve_resource("2301.07041").await?;
 
   dbg!(&paper);
   // assert!(resource.validate(&paper)?);
@@ -75,10 +81,11 @@ async fn test_iacr_retriever_integration() -> TestResult<()> {
   );
 
   let retriever: Retriever = toml::from_str(&ret_config_str).expect("Failed to parse config");
-  let resource: ResourceConfig = toml::from_str(&res_config_str).expect("Failed to parse config");
+  // let resource: ResourceTemplate = toml::from_str(&res_config_str).expect("Failed to parse
+  // config");
 
   // // Test with a real IACR paper
-  let paper = retriever.retrieve_resource("2016/260", &resource).await.unwrap();
+  let paper = retriever.retrieve_resource("2016/260").await.unwrap();
   // assert!(resource.validate(&paper)?); // TODO: validation already happens internally, to be fair
   // that validation may not be working totally right
   dbg!(&paper);
@@ -128,10 +135,10 @@ async fn test_doi_retriever_integration() -> TestResult<()> {
   );
 
   let retriever: Retriever = toml::from_str(&ret_config_str).expect("Failed to parse config");
-  let resource: ResourceConfig = toml::from_str(&res_config_str).expect("Failed to parse config");
+  let resource: ResourceTemplate = toml::from_str(&res_config_str).expect("Failed to parse config");
 
   // Test with a real DOI paper
-  let paper = retriever.retrieve_resource("10.1145/1327452.1327492", &resource).await?;
+  let paper = retriever.retrieve_resource("10.1145/1327452.1327492").await?;
   // assert!(resource.validate(&paper)?);
   dbg!(&paper);
   // assert!(!paper.title.is_empty());
